Move redis client constructor to cache package
