+++ api/internal/cache/v1/pkg/remote/redis_cache_test.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	cachev1 "github.com/fjarm/fjarm/api/internal/cache"
 	"github.com/fjarm/fjarm/api/internal/logkeys"
-	"github.com/fjarm/fjarm/api/internal/redis/v1/pkg/client"
 	"github.com/redis/rueidis"
 	"github.com/testcontainers/testcontainers-go/modules/redis"
 	"io"
@@ -44,7 +43,7 @@ func TestMain(m *testing.M) {
 	}
 
 	addrs := rueidis.MustParseURL(connectionURI).InitAddress
-	rdb, err = client.NewRedisClient(nil, rueidis.AuthCredentials{}, addrs)
+	rdb, err = NewRedisClient(nil, rueidis.AuthCredentials{}, addrs)
 	if err != nil {
 		slog.Error("failed to create Redis client", slog.Any(logkeys.Err, err))
 	}
