+++ api/internal/cache/v1/pkg/remote/redis_client.go
@@ -1,4 +1,4 @@
-package client
+package remote
 
 import (
 	"crypto/tls"
